requirements.txt
pyyaml
flask
htmlmin2
pytest
icecream
black